added styling to Date and changed file name
