+++ eq-author/src/components/AdditionalContent/AnswerProperties/Answers/index.js
@@ -1,5 +1,5 @@
 export { default as Decimal } from "../Decimal";
-export { default as DateFormat } from "./DateFormat";
+export { default as DateProperties } from "./DateProperties";
 export { default as UnitProperties } from "./UnitProperties";
 export { default as DurationProperties } from "./DurationProperties";
 export { default as TextProperties } from "./TextProperties";
